+++ CHANGES.md
@@ -3,7 +3,14 @@
 -   Update documentation to eliminate references to `this.state`,
     as this can lead to odd bugs.
 
--   Also export exportables from ``state`` module, like FormState.
+-   Also export exportables from `state` module, like FormState.
+
+-   SubForm support. You can get fields for a nested object by using `SubForm`,
+    analogous to how you can fields for an array using `RepeatingForm`.
+
+-   Accessors previously had a `node` property. I realized I could remove this
+    as we always can reconstruct the node from the path where needed
+    within the accessors.
 
 # 0.12.1